requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-environ>=0.11,<1.0
django-filter>=24.0
django-cors-headers>=4.4
psycopg2-binary>=2.9
Pillow>=10.0
celery>=5.4
redis>=5.0
anthropic>=0.40.0
twilio>=9.0
